+++ wetterdienst/ui/explorer/layout/main.py
@@ -5,8 +5,7 @@
 Wetterdienst Explorer UI layout.
 """
 import dash_bootstrap_components as dbc
-import dash_core_components as dcc
-import dash_html_components as html
+from dash import dcc, html
 
 from wetterdienst.ui.explorer.layout.observations_germany import dashboard_layout