+++ wetterdienst/ui/explorer/library.py
@@ -20,26 +20,12 @@ def default_figure(climate_data: pd.DataFrame, column: str) -> go.Figure:
     Returns:
         plotly Figure object
     """
-
-    # Create a placeholder figure for empty data.
-    # https://community.plotly.com/t/replacing-an-empty-graph-with-a-message/31497/2
-    if column is None:
-        fig = go.Figure()
-        fig.add_annotation(
-            text="No variable selected",
-            xref="paper",
-            yref="paper",
-            showarrow=False,
-            font={"size": 28},
-        )
-        return fig
-
     if climate_data.empty:
         fig = go.Figure()
         add_annotation_no_data(fig)
         return fig
 
-    fig = go.Figure(data=[go.Scatter(x=climate_data.date, y=climate_data.loc[:, column], hoverinfo="x+y")])
+    fig = go.Figure(data=[go.Scatter(x=climate_data.date, y=climate_data.value, hoverinfo="x+y")])
     fig.update_layout(yaxis={"title": f"{column}"}, showlegend=False)
     return fig